Start splitting pgx into separate ingest and reader files
These paths have little to do with each other, besides for both talking
to the database. By splitting them we should reduce merge conflicts.
